+++ sample_project/cfehome/views.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
-from django.http import HttpResponse
 from django.shortcuts import render
 
 
 def index(request):
-    return render(request, 'index.html')
+    return render(request, "index.html")